tests/submission/Opera/getElementsByClassName/009.htm

<!doctype html>
<html class="a A">
 <head>
  <title>document.getElementsByClassName(): case sensitive</title>
  <script src="../resources/testharness.js"></script>
 </head>
 <body class="a a">
  <div id="log"></div>
  <script>test(function() {
                 assert_array_equals(document.getElementsByClassName("A a"), [document.documentElement])
               }) 
  </script>
 </body>
</html>
